+++ jobbergate-cli/CHANGELOG.md
@@ -4,8 +4,10 @@ This file keeps track of all notable changes to jobbergate-cli
 
 ## Unreleased
 
+- Modified the Question/Answer workflow to make it behave like jobbergate-legacy [ASP-4332]
 
 ## 4.4.0a0 -- 2024-02-19
+
 - Added command to render an application into a job script locally [ASP-3326]
 
 ## 4.3.0 -- 2024-02-14